user-service/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
bcrypt==4.3.0
python-jose[cryptography]
python-multipart
pydantic
pydantic
pydantic[email]
